refactor(ios, webview): extract WebInteraction.
