+++ detox/ios/Detox/Invocation/WebAction.swift
@@ -3,31 +3,20 @@
 //  Created by Asaf Korem (Wix.com) on 2024.
 //
 
-import Foundation
 import WebKit
 
-class WebAction: CustomStringConvertible {
-	var predicate: Predicate?
-	var atIndex: Int?
-	var webPredicate: WebPredicate
+class WebAction: WebInteraction {
 	var webAction: WebActionType
 	var params: [Any]?
 
-	init(json: [String: Any]) throws {
+	override init(json: [String: Any]) throws {
 		self.webAction = WebActionType(rawValue: json["webAction"] as! String)!
-
 		self.params = json["params"] as? [Any]
+		try super.init(json: json)
+	}
 
-		let webPredicateJSON = json["webPredicate"] as! [String: Any]
-		let webPredicateData = try JSONSerialization.data(withJSONObject: webPredicateJSON)
-		let decodedWebPredicate = try JSONDecoder().decode(WebPredicate.self, from: webPredicateData)
-
-		self.webPredicate = decodedWebPredicate
-		if let predicateJSON = json["predicate"] as? [String: Any] {
-			self.predicate = try Predicate.with(dictionaryRepresentation: predicateJSON)
-		}
-
-		self.atIndex = json["atIndex"] as? Int
+	override var description: String {
+		return "\(webAction.rawValue)"
 	}
 
 	func perform(completionHandler: @escaping ([String: Any]?, Error?) -> Void) {
@@ -76,26 +65,5 @@ class WebAction: CustomStringConvertible {
 		} catch {
 			completionHandler(nil, error)
 		}
-
 	}
-
-	var description: String {
-		return "\(webAction.rawValue)"
-	}
-}
-
-enum WebActionType: String, Codable {
-	case tap = "tap"
-	case typeText = "typeText"
-	case replaceText = "replaceText"
-	case clearText = "clearText"
-	case selectAllText = "selectAllText"
-	case getText = "getText"
-	case scrollToView = "scrollToView"
-	case focus = "focus"
-	case moveCursorToEnd = "moveCursorToEnd"
-	case runScript = "runScript"
-	case runScriptWithArgs = "runScriptWithArgs"
-	case getCurrentUrl = "getCurrentUrl"
-	case getTitle = "getTitle"
 }
