+++ detox/ios/Detox/Invocation/WebExpectationModifier.swift
@@ -1,0 +1,8 @@
+//
+//  WebExpectationModifier.swift (Detox)
+//  Created by Asaf Korem (Wix.com) on 2024.
+//
+
+enum WebModifier: String, Codable {
+	case not = "not"
+}
