+++ detox/ios/Detox/Invocation/WebExpectationType.swift
@@ -1,0 +1,9 @@
+//
+//  WebExpectationType.swift (Detox)
+//  Created by Asaf Korem (Wix.com) on 2024.
+//
+
+enum WebExpectationType: String, Codable {
+	case toExist = "toExist"
+	case toHaveText = "toHaveText"
+}
